+++ src/core/calendar/CommunityCollegeAfterwardsWorkingHolidayCalendar/generate.ts
@@ -1,0 +1,27 @@
+import { AGE_OF_START_STORY } from "../../../constants/fullcalendar/options";
+import { WORKING_HOLIDAY_APPLICATION_LIMITATION_AGE } from "../../../constants/visa";
+import { COMMUNITY_COLLEGE_EXAMPLE1 } from "../../../constants/school";
+import { range } from "../../../lib/util";
+import { addYears, addMonths } from "date-fns";
+import { build } from "../../../core/story/CommunityCollegeAfterwardsWorkingHolidayStory/CommunityCollegeAfterwardsWorkingHolidayStory";
+
+const startMonths = COMMUNITY_COLLEGE_EXAMPLE1.startMonths;
+
+const addingNumbers = range(
+  AGE_OF_START_STORY,
+  WORKING_HOLIDAY_APPLICATION_LIMITATION_AGE
+);
+
+export const generate = (birth: Date) => {
+  return addingNumbers
+    .map((num) => addYears(birth, num))
+    .flatMap((startDate) => {
+      const datesInYear = startMonths.map((month) =>
+        addMonths(startDate, month)
+      );
+      return datesInYear;
+    })
+    .map((startDate) => {
+      return build({ startDate });
+    });
+};
